Part 1 of the great clone plan.
When add_method is called with a method object, it calls clone on that method.

To facilitate tracking the source of a method, we save the original
method (the clone source) in the newly cloned method.

I also added a bunch of convenience methods for getting various
original names out of said original method. Eventually this will be
used in Moose to determine whether a method originally came from a
role.