Fix RT 48985
When add_method is called, it now always adds _something_ to the method map,
but it may be a raw coderef. In get_method, we check for this and inflate it
into a proper method meta-object as needed.

This means that the method map contains a mix of objects and code refs, and so
is no longer fit for public consumption.